+++ integration-tests/src/test/java/arez/integration/dagger/package_access/other/TestComponent.java
@@ -3,11 +3,10 @@ package arez.integration.dagger.package_access.other;
 
 import arez.annotations.ArezComponent;
 import arez.annotations.Feature;
-import arez.annotations.InjectMode;
 import arez.annotations.Observable;
 import arez.annotations.PostConstruct;
 
-@ArezComponent( dagger = Feature.ENABLE, inject = InjectMode.PROVIDE )
+@ArezComponent( dagger = Feature.ENABLE )
 public abstract class TestComponent
 {
   TestComponent( final MyDependency myDependency )