+++ processor/src/test/fixtures/bad_input/com/example/inject/FieldInjectionModel.java
@@ -2,10 +2,9 @@ package com.example.inject;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.InjectMode;
 import javax.inject.Inject;
 
-@ArezComponent( inject = InjectMode.NONE )
+@ArezComponent
 public abstract class FieldInjectionModel
 {
   @Inject