+++ processor/src/test/fixtures/input/com/example/inject/NoInjectModel.java
@@ -2,9 +2,9 @@ package com.example.inject;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.InjectMode;
+import arez.annotations.Feature;
 
-@ArezComponent( inject = InjectMode.NONE )
+@ArezComponent( dagger = Feature.DISABLE )
 public abstract class NoInjectModel
 {
   @Action