+++ processor/src/test/fixtures/input/com/example/inject/PublicCtorDaggerModel.java
@@ -1,0 +1,12 @@
+package com.example.inject;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+
+@ArezComponent( service = Feature.ENABLE, allowEmpty = true )
+public abstract class PublicCtorDaggerModel
+{
+  public PublicCtorDaggerModel()
+  {
+  }
+}