+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Remove `arez.annotations.InjectMode` and related infrastructure as it is unused.
+
 ### [v0.168](https://github.com/arez/arez/tree/v0.168) (2020-02-16) · [Full Changelog](https://github.com/arez/arez/compare/v0.167...v0.168) · [API Differences](https://arez.github.io/api-diff?key=arez&old=0.167&new=0.168)
 
 The release includes 3 non breaking API changes and 3 breaking API changes.