+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Generate a suppressable warning when a public constructor exists in an injectable type rather than a hard failure.
 * Generate an error if a constructor parameter is a parameterized type and dagger injection is enabled.
 * Generate an error if the component is a parameterized type and dagger injection is enabled.
+* Fix several bugs that could manifest if `@ArezComponent` is attached to an interface.
 
 ### [v0.168](https://github.com/arez/arez/tree/v0.168) (2020-02-16) · [Full Changelog](https://github.com/arez/arez/compare/v0.167...v0.168) · [API Differences](https://arez.github.io/api-diff?key=arez&old=0.167&new=0.168)